cloud-deploy/backend/requirements.txt
# 业主大会投票小程序 - 后端依赖

# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9          # PostgreSQL（本地开发）
pymysql==1.1.0                    # MySQL（云托管）
cryptography==42.0.0              # MySQL 加密连接支持
alembic==1.13.1

# 缓存
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0
pyjwt==2.8.0

# OCR
baidu-aip==4.16.13

# 区块链
requests==2.31.0

# 消息推送
wechatpy==2.0.0a26

# 工具
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# 日志
loguru==0.7.2
